Fix warnings

+++ src/UpdraftPkg/Updraft/Graphics/Point.cpp
@@ -18,7 +18,7 @@ void Point::draw(const Color color) const
 {
   const uint32 width = Screen::Width(), height = Screen::Height();
 
-  if (x < 0 || x >= width || y < 0 || y >= height)
+  if (x < 0 || x >= static_cast<int32>(width) || y < 0 || y >= static_cast<int32>(height))
     return;
 
   auto *frameBufferBase = Graphics::BltBuffer();
